removed id from createViolationRecordSchema
