+++ src/modules/violationRecord/src/dtos/violationRecordRequestSchema.ts
@@ -2,7 +2,6 @@ import z from "zod";
 import { ViolationRecordStatus } from "@prisma/client"; 
 
 export const ViolationRecordSchema = z.object({
-  id: z.string().uuid(),
   userId: z.string().uuid(),
   reportedById: z.string().uuid(),
   violationId: z.string().uuid(),
